added code and test
